[GLST] worked on stopping Threading issues (all swing runs directly on event dispatching thread or on that thread via SwingUtilities.invokeLater) and making GC.startGame interruptible.[GLST] built in enable/disable for the start game button, and made button work.  BUG: IF CLIENT START'S GAME BEFORE HOST, DEADLOCK RESULTS.  NEEDS TO BE ADDRESSED.
On the Galaxy Designer side, I have finally made foci selectable and draggable.  To this end, the empty interface Selectable has been made.  Focus and StellarObject implement it.  orbit.focus2_x/focus2_y have been replaced with a Focus object.

Evidence now suggests Orbit.calcTimeOffset is the culprit for the incorrect init position bug.  That function needs to be reworked.

Naming Galaxies is now possible, and, in fact, forced by the Galaxy Designer.

I also experimented with custom cursors in SystemViewer and have implemented them for moving the camera around when the mouse is near the edge of the viewer.  Also, the camera motion is now on a short delay before it kicks in.

Bugs fixed in the Galaxy view in the Galaxy Designer - particularly, shift and alt-clicking now toggles a system from selected to unselected, as it should.  The bug was due to that shift_down_on_click and alt_down_on_click were only set under certain conditions in mousePressed but should have been set no matter what.  This has been fixed.


git-svn-id: https://subversion.ews.illinois.edu/svn/GameBuilders/GalacticStrategy@1042 e605e84e-813e-0410-88cd-b83e9fd653cd