[GLST] finished implementing Warping - current behavior is workable, but just a draft.  Some small issues need to be smoothed out, such as the ability to cancel a warp after clicking the warp button, and ability to see destination before clicking.
Warp implementation has come along with a number of new interface bugs.  For one, the ShipCommandPanel continues to display the old destination, and the SystemPainter continues to draw the old red cross to denote where the old destination was.

Also, there seems to be an issue with the cost JLabel when switching between systems.  Also, ConcurrentModificationException was observed with GSystem.fleets - I believe I have obviated the problem with Fleet's new member lock.  add and remove are now safe, but any iteration over the Fleet's ships must be synchronized on the Fleet's lock object.

The worst bug right now, however, is that ships can shoot themselves... oops =0.  Should be fixed soon.

git-svn-id: https://subversion.ews.illinois.edu/svn/GameBuilders/GalacticStrategy@1977 e605e84e-813e-0410-88cd-b83e9fd653cd