[GLST] numerous coordination fixes, restricting various actions.  Fixed pickup troops feature.  I think it is still possible to order an enemy ship to warp - but the interface displayed when looking at an enemy ship needs to change anyway.
Fixed infinite recursion of ModifiedBinarySearch in RelaxedDataSaverControl.

Rewrote GalacticMapPainter to use Antialiased graphics.  The only issue is that the triangle used to represent a warping ship seems to warp as it moves along - probably because java attempts to prevent antialiasing from creating a "twice as thick half as opaque" effect.

Changed ship images to be loaded as BufferedImages instead of plain Images.  As a result, ImageSizer is unnecessary.  Also, modified how ships are drawn so that they are now antialiased and use double coordinates :).  As part of this, I added caching of the scaled image of the ship in ShipType.  Also saving the ImageIcon produced from a ship image since it never changes but before would be created several times.

git-svn-id: https://subversion.ews.illinois.edu/svn/GameBuilders/GalacticStrategy@2103 e605e84e-813e-0410-88cd-b83e9fd653cd