Major code quality upgrade - removed most timing variables.  The ones that now exist are passed as function parameters; no more worrying about keeping times in sync across all objects.  As a result, some functions now need the time value (no time on the object to use), but most functions lost time as a parameter.
The other major timing change is that I'm now forcing all orders onto a time grain.

Was able to remove the sanity checks in DataSaver - no longer necessary because the times to keep in sync don't exist anymore.

Adding test cases 12 and 13.  Also bashed a bug related to soldiers, resulting in test case 13 working, as well as a test startup problem related to most_recent_time in GameUpdater (I believe test 12 hit trouble with this).